+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.1.0
 Initial release (development still in progress)
 
+
+## 0.2.0
+Add - #2 Create Pester scripts
+
+Add - #6 Display somehow progression of Startup Scripts
+
+Add - #9 Add WithoutStarting swith to Install-PowerLSS.ps1
+
+Fix - #10 Prevent from computer restart loop